requirements.txt
redis
imgurpython
sqlalchemy
python-telegram-bot==12.2.0
Pillow
psycopg2
